Fix output documentation
`Output` is a thing the user defines, not a built-in Cog thing. I'm not sure the best hierarchy for this in the document, but this is clearer than it was before.

I have removed the API documentation because this is confusing for people trying to implement models for Replicate. Perhaps this should live in some separate API documentation?