Pick latest patch version for CUDA
This worked before because for 11.1, 11.1.0 existed, and this all
accidentally worked because the empty value for an int in Go is 0.

But, 11.1.0 doesn't seem to exist any longer so that all broke because
"11.1" is not equal to "11.1.1".

This also actually didn't work properly, because there is a later
version of CUDA 11.1 that was never picked.

This all works properly now, by resolving to the latest minor version of
CUDA.

Signed-off-by: Ben Firshman <ben@firshman.co.uk>