Update compatibility matrices (#504)
* Relax generate compatibility to handle ROCm

"RocM" is now "ROCm"

Signed-off-by: Ben Firshman <ben@firshman.co.uk>

* More forgiving parser to deal with --extra-index-url

Torch 1.11 uses this option in the install command for CUDA 11.3. For
CUDA 10.2 it has no options and falls back to PyPi.

But, what's weird, is both Torch's custom index and PyPi have exactly
the same packages (I checked locally with SHA sum) so this option isn't
even used. In other words, CPU, CUDA 10.2, and CUDA 11.3 are now all
just the standard PyPi library

I was considering passing through --extra-index-url through to our
install, but gave up because it seems redundant. This may change in the
future, or might just be a bug in PyTorch's release.

Another minor change: PyTorch's GPU packages don't include versions now,
like the CPU packages previously.

Signed-off-by: Ben Firshman <ben@firshman.co.uk>

* Fetch latest torch versions from PyPi

They're not specified in the `pip install` string on pytorch.org any
longer.

Signed-off-by: Ben Firshman <ben@firshman.co.uk>

* Add Python 3.10 support for PyTorch

We don't actually do Python version checking for PyTorch yet, so just
add support for everything

Signed-off-by: Ben Firshman <ben@firshman.co.uk>

* Pick latest patch version for CUDA

This worked before because for 11.1, 11.1.0 existed, and this all
accidentally worked because the empty value for an int in Go is 0.

But, 11.1.0 doesn't seem to exist any longer so that all broke because
"11.1" is not equal to "11.1.1".

This also actually didn't work properly, because there is a later
version of CUDA 11.1 that was never picked.

This all works properly now, by resolving to the latest minor version of
CUDA.

Signed-off-by: Ben Firshman <ben@firshman.co.uk>

* Update compatibility matrices

Signed-off-by: Ben Firshman <ben@firshman.co.uk>