Don't expect prediction ID from queueing API request
We were previously using this to tag logs, sent to a separate logging
queue. Now that logs are included in the response we don't use this
value for anything meaningful.

Dropping it from the expected request format takes us slightly closer to
the HTTP API format, which doesn't have this.

The only remaining difference now should be the response queue.

Signed-off-by: Dominic Baggott <dominic.baggott@gmail.com>