Use a string key for responses instead of a list
We're sending complete snapshots, so we don't need to keep intermediate
states. The downside of putting the complete states onto a regular list
is that we're adding monotonically increasing snapshots to it. In
extreme cases, the size and number of those items can overwhelm the
Redis server.

Using a string key should also be slightly more resilient than popping
off a list, as we don't lose the latest state if something goes wrong.

This one line change requires a significant update to the tests. Using
`GET` instead of `BLPOP` introduces some timing issues, where the Cog
predictor under test can `SET` the key value twice in quick succession,
before we have a chance to read it.

Where possible I've introduced short sleeps into the predictors, to
allow us to read the outputs in sequence. Where that's not possible (for
example, the quick-fire responses for a traceback in logs followed by a
failed response) we check that two responses come in, without checking
the content of the first.

It's possible that there are still timing issues in these tests, but if
so they'll only come to light after repeated runs in different
environments.

Signed-off-by: Dominic Baggott <dominic.baggott@gmail.com>