Switch to keyword arguments for queue worker
The positional arguments are trouble waiting to happen. Switch over to
keyword arguments which are less fragile, and can provide typing
automatically.

We'll need to do at least one release with this before dropping
positional arguments, to give folks a chance to update how they run Cog.
We we do drop positional arguments we should make most of the keyword
arguments required.

I've updated a couple of tests but left the rest, so we can be confident
both positional and keyword arguments work. We can swap the rest when we
remove positional args.

Signed-off-by: Dominic Baggott <dominic.baggott@gmail.com>