Make File inputs survive pickling/unpickling
This commit makes it possible for the HTTP API to handle File input
types with the new Worker. Because Worker is on the other side of a
process boundary, whatever the inputs validate to must survive a
pickle/unpickle, and file handles will not ever meet this criterion.

This addresses the problem for File, by mapping URLs to URLFile, a
"file-like" object that lazily executes the HTTP request when it is
first used. This works, but does have some downsides:

- File and Path now behave differently: File inputs are downloaded in
  the predictor process, Path inputs are downloaded at validation-time.

- We now don't detect problems with the URL (e.g. it 404s) until we
  actually call predict().

Signed-off-by: Nick Stenning <nick@whiteink.com>