Preserve prediction status after slow cancelation
By calling `tracker.fail()` we overrode the `canceled` status to be
`failed`, which has unintended side effects and leaks the internals of
the system -- it doesn't matter to an end user whether we cleanly
termianted the prediction or not, just that it was ended. Either way it
ended up canceled.

If we're here because the prediction timed out this is even more
misleading, as the error message is about being canceled but the reason
is the prediction timed out.

Signed-off-by: Dominic Baggott <dominic.baggott@gmail.com>