Unpin library deps
It doesn't seem wise to pin our primary dependencies, as the consumers
of the `cog` package will not have them pinned, and the exact versions
selected by `pip` will depend on the version of Python under which the
installation occurs as well as any other platform or environmental
version selectors specified by our dependencies.

This unpins the library dependencies so we're testing Cog in the same
environment that it will be used.

Signed-off-by: Nick Stenning <nick@whiteink.com>